+++ packages/react-libraries/CHANGELOG.md
@@ -15,6 +15,12 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Breaking changes
+
+- Don't replace `process.env.NODE_END` in CommonJS and ES modules bundles.
+
 ## 0.5.0 (July 10, 2018)
 
 ### Enhancements